requirements.txt
astroid==2.5.6
beautifulsoup4==4.9.3
breadability==0.1.20
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
cssselect==1.1.0
docopt==0.6.2
feedfinder2==0.0.4
feedparser==6.0.2
filelock==3.0.12
idna==2.10
isort==5.8.0
jieba3k==0.35.1
joblib==1.0.1
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.20.3
Pillow==8.2.0
pycountry==20.7.3
pylint==2.8.2
python-dateutil==2.8.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
sumy==0.8.1
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
tqdm==4.61.0
urllib3==1.26.5
Wand==0.6.6
wrapt==1.12.1
